Added batch training, a batch object, updated progress bar, and added matrix operations, along with a matrix utils class

+++ include/core/Data.h
@@ -10,9 +10,9 @@ class Data {
     private:
         // Instances Variables
         vector<vector<double> > trainFeatures;
-        vector<double> trainTarget;
+        vector<int> trainTarget;
         vector<vector<double> > testFeatures;
-        vector<double> testTarget;
+        vector<int> testTarget;
         bool isDataLoaded;
         static const double MAX_GREYSCALE_VALUE;
 
@@ -20,15 +20,14 @@ class Data {
         static random_device rd;
         static mt19937 generator;
 
-
         // Private Methods
         void readData(string, bool, int);
         void minmaxData();
         void minmaxNormalizeColumn(vector<vector<double> >&, double, double, int);
         void getMinMaxColumn(vector<vector<double> >&, double&, double&, int);
         void checkFile(string);
-        void parseLine(string, vector<vector<double> >&, vector<double>&, int);
-        void setData(vector<vector<double> >&, vector<double>&, bool);
+        void parseLine(string, vector<vector<double> >&, vector<int>&, int);
+        void setData(vector<vector<double> >&, vector<int>&, bool);
         void normalizeGreyScale(vector<vector<double> >&); 
     
     public:
@@ -38,8 +37,8 @@ class Data {
         // void readAllData(string, int, float);
         const vector<vector<double> >& getTrainFeatures() const;
         const vector<vector<double> >& getTestFeatures() const;
-        const vector<double>& getTrainTarget() const;
-        const vector<double>& getTestTarget() const;
+        const vector<int>& getTrainTarget() const;
+        const vector<int>& getTestTarget() const;
         void minmax();
         void minmaxGreyScale();
         int getTrainFeatureSize() const;
